Add EditMode for details

+++ attendance_system_frontend/src/app/employee-details/employee-details.component.ts
@@ -1,14 +1,15 @@
 import { Component, OnInit } from '@angular/core';
 import { ActivatedRoute, Params, Router } from '@angular/router';
+import { FormsModule, NgForm } from '@angular/forms';
 
 import { EmployeeService } from '../services/employee.service';
 import { DepartmentService } from '../services/department.service';
-import { Department, Employee } from '../models/app-models';
+import { Department, Employee, Gender, Position, UserRole } from '../models/app-models';
 
 @Component({
   selector: 'app-employee-details',
   standalone: true,
-  imports: [],
+  imports: [FormsModule],
   templateUrl: './employee-details.component.html',
   styleUrl: './employee-details.component.css'
 })
@@ -17,12 +18,26 @@ export class EmployeeDetailsComponent implements OnInit {
   selectedEmployee: Employee;
   selectedEmployeeId: number;
 
+  updatedEmployee: Employee;
+
   position: string;
   gender: string;
   userRole: string;
   department: string;
+  currentHireDate: string;
 
   departments: Department[];
+
+  editMode: boolean = false;
+
+  genders = Object.keys(Gender).filter(key => isNaN(Number(key)));
+  gendersValue = Object.values(Gender).filter(value => !isNaN(Number(value)));
+
+  positions = Object.keys(Position).filter(key => isNaN(Number(key)));
+  positionsValue = Object.values(Position).filter(value => !isNaN(Number(value)));
+
+  userRoles = Object.keys(UserRole).filter(key => isNaN(Number(key)));
+  userRolesValue = Object.values(UserRole).filter(value => !isNaN(Number(value)));
 
   constructor(private route: ActivatedRoute,
     private router: Router,
@@ -34,12 +49,17 @@ export class EmployeeDetailsComponent implements OnInit {
       this.departments = departments;
     });
 
+    this.onSelectedEmployee();
+  }
+
+  public onSelectedEmployee() {
     this.route.params.subscribe((params: Params) => {
       this.selectedEmployeeId = +params['id'];
 
       if (this.selectedEmployeeId) {
         this.employeeService.getEmployeeById(this.selectedEmployeeId).subscribe((employee) => {
           this.selectedEmployee = employee;
+          this.currentHireDate = new Date(this.selectedEmployee.hireDate).toISOString().split('T')[0];
           this.setPosition();
           this.setGender();
           this.setUserRole();
@@ -92,16 +112,64 @@ export class EmployeeDetailsComponent implements OnInit {
   }
 
   public OnDeleteEmployee(id: number) {
-    this.employeeService.deleteEmployee(id).subscribe({
-      next: () => {
-        alert('Employee deleted successfully!');
-        this.onEmployeeList();
-      },
-      error: (err) => {
-        console.log('Delete error: ',err);
-        alert('Failed to delete employee.');
-      }
-    });
+    if (window.confirm('Are you sure you want to delete this employee?')) {
+      this.employeeService.deleteEmployee(id).subscribe({
+        next: () => {
+          this.onEmployeeList();
+        }
+      });
+    }
+  }
+
+  public onEditMode() {
+    this.onUpdateEmployee();
+    this.editMode = true;
+  }
+
+  public onSubmit(detailsEditMode: NgForm) {
+    if (detailsEditMode.invalid) {
+      alert('Please fill out all required fields.');
+    }
+    else {
+      this.employeeService.updateEmployee(this.selectedEmployeeId, this.selectedEmployee).subscribe({
+        next: (employee) => {
+          alert('Employee updated successfully!');
+          this.onSelectedEmployee();
+          this.editMode = false;
+        },
+        error: (err) => {
+          alert('Failed to update employee!');
+        }
+      });
+    }
+  }
+
+  onCancelForm() {
+    if (window.confirm('Are you sure you want to cancel?')) {
+      this.onSelectedEmployee();
+      this.editMode = false;
+    }
+  }
+
+  onUpdateEmployee() {
+    this.updatedEmployee = { ...this.selectedEmployee };
+    this.updatedEmployee.gender = this.selectedEmployee.gender;
+    this.updatedEmployee.position = this.selectedEmployee.position;
+    this.updatedEmployee.userInfoDTO = { ...this.selectedEmployee.userInfoDTO };
+    this.updatedEmployee.departmentId = this.selectedEmployee.departmentId;
+    this.updatedEmployee.addressDTO = { ...this.selectedEmployee.addressDTO };
+    this.updatedEmployee.attendanceRecordDTOs = { ...this.selectedEmployee.attendanceRecordDTOs };
+  }
+
+  onUpdateSelectedEmployee() {
+
+    this.selectedEmployee = { ...this.updatedEmployee };
+    this.selectedEmployee.gender = this.updatedEmployee.gender;
+    this.selectedEmployee.position = this.updatedEmployee.position;
+    this.selectedEmployee.userInfoDTO = { ...this.updatedEmployee.userInfoDTO };
+    this.selectedEmployee.departmentId = this.updatedEmployee.departmentId;
+    this.selectedEmployee.addressDTO = { ...this.updatedEmployee.addressDTO };
+    this.selectedEmployee.attendanceRecordDTOs = { ...this.updatedEmployee.attendanceRecordDTOs };
   }
 }
